Added AsyncWrapper

+++ examples/poll-tutorial/src/main/java/io/datakernel/examples/PollLauncher.java
@@ -14,10 +14,12 @@ import io.datakernel.http.HttpResponse;
 import io.datakernel.http.RoutingServlet;
 import io.datakernel.launcher.Launcher;
 import io.datakernel.launchers.http.HttpServerLauncher;
+import io.datakernel.writer.ByteBufWriter;
 
 import java.util.Collection;
 import java.util.Map;
 
+import static io.datakernel.http.AsyncServletWrapper.loadBody;
 import static io.datakernel.http.HttpHeaders.REFERER;
 import static io.datakernel.http.HttpMethod.GET;
 import static io.datakernel.http.HttpMethod.POST;
@@ -46,9 +48,9 @@ public final class PollLauncher extends HttpServerLauncher {
 			@Provides
 			@Singleton
 			AsyncServlet servlet(PollDao pollDao) {
-				Mustache singlePollView = new DefaultMustacheFactory().compile("site/singlePollView.html");
-				Mustache singlePollCreate = new DefaultMustacheFactory().compile("site/singlePollCreate.html");
-				Mustache listPolls = new DefaultMustacheFactory().compile("site/listPolls.html");
+				Mustache singlePollView = new DefaultMustacheFactory().compile("templates/singlePollView.html");
+				Mustache singlePollCreate = new DefaultMustacheFactory().compile("templates/singlePollCreate.html");
+				Mustache listPolls = new DefaultMustacheFactory().compile("templates/listPolls.html");
 
 				return RoutingServlet.create()
 						//[START REGION_1]
@@ -68,10 +70,14 @@ public final class PollLauncher extends HttpServerLauncher {
 						//[END REGION_2]
 						.with(GET, "/create", request -> Promise.of(HttpResponse.ok200()
 								.withBody(applyTemplate(singlePollCreate, ImmutableMap.of()))))
-						.with(POST, "/vote", request -> request.getPostParameters()
-								.map(params -> {
+						.with(POST, "/vote", loadBody()
+								.then(request -> {
+									Map<String, String> params = request.getPostParameters();
 									String option = params.get("option");
 									String stringId = params.get("id");
+									if (option ==null || stringId == null) {
+										return Promise.of(HttpResponse.ofCode(401));
+									}
 
 									int id = Integer.parseInt(stringId);
 									PollDao.Poll question = pollDao.find(id);
@@ -79,10 +85,11 @@ public final class PollLauncher extends HttpServerLauncher {
 									question.vote(option);
 
 									String referer = request.getHeader(REFERER);
-									return HttpResponse.redirect302(referer != null ? referer : "/");
+									return Promise.of(HttpResponse.redirect302(referer != null ? referer : "/"));
 								}))
-						.with(POST, "/add", request -> request.getPostParameters()
-								.map(params -> {
+						.with(POST, "/add", loadBody()
+								.then(request -> {
+									Map<String, String> params = request.getPostParameters();
 									String title = params.get("title");
 									String message = params.get("message");
 
@@ -90,14 +97,18 @@ public final class PollLauncher extends HttpServerLauncher {
 									String option2 = params.get("option2");
 
 									int id = pollDao.add(new PollDao.Poll(title, message, list(option1, option2)));
-									return HttpResponse.redirect302("poll/" + id);
+									return Promise.of(HttpResponse.redirect302("poll/" + id));
 								}))
-						.with(POST, "/delete", request -> request.getPostParameters()
-								.map(params -> {
+						.with(POST, "/delete", loadBody()
+								.then(request -> {
+									Map<String, String> params = request.getPostParameters();
 									String id = params.get("id");
+									if (id == null) {
+										return Promise.of(HttpResponse.ofCode(401));
+									}
 									pollDao.remove(Integer.parseInt(id));
 
-									return HttpResponse.redirect302("/");
+									return Promise.of(HttpResponse.redirect302("/"));
 								}));
 			}
 		});
